issue: PHP 7.2 Ticket Status
This addresses issue 4716 where updating a Ticket Status throws a fatal
error of "Too few arguments" when using PHP 7.2. This is due to `$errors`
not being passed to the `update()` function causing PHP 7.2 to freak out.
This creates an `$errors` array and passes it to `update()` so PHP 7.2 is
happy and we are all happy. :)