Issue: Ticket Export Headers
This commit fixes an issue where custom export headings were being set to numeric values rather than field labels.

For exports that are already saved, we can get the heading directly using getHeading. To get the heading for fields that are not already saved as exports, we can use getExportableFields, which will return path => heading.