Issue: Choosing Fields to Export
This commit fixes an issue where only the saved export fields would export, whether you checked more fields to export or unchecked some of the saved fields.

For any fields not saved that need to be exported, we need to find their label and then put it in the fields array. For any saved export fields that are unchecked, we need to compare what was checked vs what is saved and then unset the fields that were unchecked.