issue: Task Print PDF
This addresses an issue where clicking print on a Task throws a fatal error.
This is due to the Task2PDF function passing a string instead of an array to
`Mpdf::__construct()`. The `__construct()` of class Mpdf requires an array
of configurations so we are now passing an array instead of a string.