feature: ACL (Access Control List)
This adds a new feature called ACL that offers the ability to control what
IP addresses are allowed to access the system. This adds a new textfield to
the System Settings to add a comma separated list of IPs. This also adds a
dropdown labeled "Apply To:" that gives you the option to choose which
panel(s) the ACL will apply to. Lastly, this adds a validator for a simple
comma-separated list of IP addresses. (eg. `192.168.1.1, 192.168.2.2,
192.168.3.3`)

If the requester's IP is not in the ACL the system will show an "Access
Denied" page. If the requester's IP is in the ACL they will be able to
access the system as usual. If the ACL field is set to Disabled, anyone will
be able to access the system.

This adds an initial failsafe where if the Admin's current IP address is not
in the ACL upon saving the system will refuse to save the setting to prevent
the Admin from being locked out. This also adds another failsafe where if
there is an "Apply To:" option set but there are no IPs provided the system
will return an error letting the Admin know they have to insert an IP
address to continue.