issue: Choice Validation Accept Punctuation
This addresses issue mentioned in 4071 where the choice field validation is
not accepting punctuations. This updates the regex to accept anything the
user inputs. This introduces another issue that if the choice contains `:`
in the value then it will truncate everything after. This is due to the
explode method that explodes on any `:` character. This updates the explode
method to return only 2 values which only explodes on one `:` character
which returns the full value.