issue: Task EstDueDate
This addresses an issue where having the `%{task.due_date}` variable in a
Task alert template whilst creating a Task with an Assignee causes a fatal
error in the backend leaving the Create Task modal spinning forever. This is
due to the `getVar()` method for Tasks where the `due_date` case references
a non-existing method. This updates the referenced method to an existing
method so it gets the correct value and does not cause a fatal error.