export: Organizations/Users Tickets
This addresses issue 4822 where exporting Tickets in an Organization or User
shows keys as values for some fields and does not populate actual data. This
is due to the exporter that uses `CustomQueue::getExportableFields()` to get
the exportable fields by default. This method does not format cdata fields
and alike properly which shows the keys instead of the actual values. This
adds a `getExportableFields()` method to class Ticket that properly formats
the cdata fields and alike which returns the appropriate data.