Issue: Ticket View Permissions for Merged Tickets
This commit changes merged Ticket permissions to the following:

If the Agent does not have access to the Child ticket and the threads are merged, they should not be allowed to view the Parent. If the threads are not merged, the Agent should still be able to access the Parent Ticket.