issue: API Validate Request Structure
This addresses an issue where validating the request structure for an API
call throws warnings when field values are arrays. This adds a new variable
called `$level` to keep track of the level of the request structure we are
in. Once we reach a certain level we will break and continue with the other
fields. This will prevent warnings when a value is an array like `["cc"] =>
{ [0] => "email2", [1] => "email1"}`.