Delete Child Ticket Issue
This commit ensures that if you choose to delete child tickets when doing a merge, we set the parent ticket's merge type to visual. Since the child ticket merged into it no longer exists, the parent ticket should act as a normal ticket that can continue to be merged as needed with any other ticket.