xss: Overridetype Broken Images
This addresses a security issue reported by Vincent Monier (Xenos) where
someone can upload a broken image containing arbitrary code, add
an overridetype to the download URL, send the link to anyone, and if clicked
will display the image as HTML instead of a broken image ergo executing the
code. This is due to an old if statement that set the file type to
the overridetype if one is provided. We don’t use this if statement anymore
so we are going to retire it. This will prevent anyone from
faking/overriding the file type which will prevent XSS in broken images. In
addition, this adds a new check with `exif_imagetype()` to ensure the file
contents match the expected file type.