issue: Form Instructions Translation
This addresses issue 4909 where setting a translated Form Instruction fails
to render properly on the client side. This is due to us formatting the
instructions but not decoding before we render. This adds the decode method
to the translated Form Instructions so they are rendered properly on the
client side.