Show Plugin if Installed and Enabled
Function getPluginByName added to class.plugin.php. This function has 2 variables that can be passed in: $name and $active (bool default false). It will return the name of a plugin if it is found in the database. If $active is true, it will also check to see if the plugin is active before returning the name of the plugin.

Using this function, we can be sure that the plugin is invisible and no extra code is run to try to log an audit if the plugin is disabled or not installed at all.