issue: Single Quotes In DB Password
This addresses an issue reported on twitter where single quotes in a db
password causes 500 Error after installation. This is due to the system not
escaping the single quotes before writing to the config file. This adds
`addcslashes()` to the password before it’s saved to the config file so it
escapes any single quotes properly.