config: Add option to detect and flag spam email
If a message positively fails the SPF and DKIM tests, then it is
flagged as spam in the thread and automated correspondance is disabled.
Because this feature may introduce delays in the processing overhead and
may not be necessary for internal help desks, the configuration option
allows for it to be disabled.