Automatic Merge Referrals
When merging tickets, if children have a different department from the parent, refer the children departments to the parent ticket. Otherwise, Agents would only be able to access the children ticket they used to have access to with a link or search because we hide merged children from the ticket queue.