issue: Advanced Search Default Sorting
This addresses an issue where conducting an Advanced Search will show the
results in ASC order (from oldest to most recent) by default. If there are
no keywords the search is supposed to show the results in DESC order (from
most recent to oldest) by default. If there are keywords the search is
supposed to show the results sorted by `relevance`. This adds additional
`select` options to the search criteria for keyword searches so that we can
use `relevance` in the outer query. This also adds an `order_by` to the
tickets to order them by the `relevance` in `DESC` order. In addition this
adds a new `order_by` to sort the results by `created` in DESC if there are
no keywords.