feature: Expanded Print View
This feature adds a new Print option labeled "Thread + Internal Notes +
Events" that, when clicked, will print the entire Ticket Thread with
Internal Notes AND all of the Thread Events. This also adds a new option in
the Ticket Print Options labeled "Print Events". Previous print options only
included the Thread only or the Thread + Internal Notes, but people want
more! This inserts the events in the correct order just like it is in the
Ticket View. The only difference between the PDF Thread Events and the
Ticket View Thread Events is the styling. In the Ticket View each event has
it's own icon to separate/differentiate them between the others. In the PDF
the events will not have their own icons rather their own background to
separate/differentiate them.