issue: DatetimeField Remove Unused Vars / Use parseDateTime()
This addresses an issue where creating a ticket with a Due Date set fails
with an error of `Select a time from the list`. This is due to unused
variables in the `Ticket::create()` function. This also improves the
datetime conversion to GMT by using `Format::parseDateTime()` instead of
trying to create a DateTime object directly which also provides a fallback
if an incorrect time. This can correctly determine the timezone (if present)
and return a proper DateTime object.