issue: THIS_VERSION Utilize MAJOR_VERSION
This addresses an issue where using osTicket from git would show the
incorrect git version in `Admin Panel > Dashboard > Information`. This is
due to a hardcoded git version of `1.8-git`. This updates `THIS_VERSION` to
utilize `MAJOR_VERSION` so that it sets the correct git version. This also
updates the `deploy` CLI module to match the same format (example:
`1.12-git`).