issue: Image Attachment View
This enhances security patch (33ed10) to allow image attachments to be
opened in a new tab if the setting is enabled. This adds a new setting to
the Agent Profile labeled "Image Attachment View" that determines how image
attachments are handled. The two options are "Download" and "Inline" with
"Download" being the default. If the option is set to "Download" all image
attachments will be forced to download. If the option is set to "Inline" all
image attachments will open in a new tab when you right click + open in new
tab.