Fix Admin Alert
This commit fixes an issue where alerts are not sent out to the Admin Email if a Department does not have any primary or extended members.

Now, if the Admin Email is checked for the New Ticket Alert and a Department does not have any members at all, an Alert will still go out to the Admin unless the Department recipients field is set to 'No one (diable Alerts and Notices)'

To make things more clear, an option has been added to have the 'Admin Only' as the recipient, so the Alert will go out ONLY to the Admin if  selected and the New Ticket Alert is enabled for the Admin Email.

If 'No one (diable Alerts and Notices)' is selected, no alert will go out at all, regardless of what is checked in Settings | Tickets | Alerts and Notices