issue: Reset Role Permissions
This addresses an issue reported on the Forum where clicking the "Reset
Permissions" action in the Agent Directory shows a blank popup. Inspecting
the console shows 500 error and checking the error logs shows `Too few
arguments to function` error. This is due to the
`RolePermission::__construct()` method expecting 1 parameter but in
`ajax.staff.php` we are not passing anything to it. This updates
`resetPermissions()` to pass `null` to `RolePermission::__construct()` so
that it doesn't complain and error out.