issue: Error On QueueSort Config
This addresses an issue where clicking on Config next to a Queue Sort will
show a blank popup. Upon viewing the error logs you receive the error `Call
to a member function asTable() on null`. This is due to defining
`$data_form` inside of the `if ($_POST)` block instead of defining it
outside the block. This updates `SearchAjaxAPI::editSort()` to define
`$data_form` outside of the `if ($_POST)` block so that `$data_form` is
always set.