Ticket Print Sort Order
This commit makes sure that the thread entries on printed tickets match the same order as the ticket in the helpdesk. A new function, sortEntries was added to class.thread.php so it can be called for the entries template as well as the ticket print template.

We also make sure that the order of thread entries display correctly for the client side tickets and printed pdfs of tickets.