issue: Dashboard No Help Error
This addresses an issue reported on the Forum where visiting the Agent
Dashboard with no Help Topics in the system completely breaks the Dashboard
view and throws a database error `You have an error in your SQL syntax;
check the manual that corresponds to your MySQL server version for the right
syntax to use near ') GROUP BY A1.topic_id' at line 1`. This is due to the
statistics code assuming we will always have `$topics` but this is not
always the case, as some people rock their helpdesk without a single Help
Topic. This adds a check to see if `$topics` is empty and if so, we return
the appropriate headers with an empty array as the plot data. This will
avoid the database error and show the appropriate Dashboard view with no
statistics listed under "Topics" tab.