Don't require refresh for inline edit
If an Agent uses inline edit to modify a single field, rather than saving the value and refreshing the page, the value will now populate in the form without having to refresh. This is useful in the event that an agent has begun writing a response to the user and does an inline edit before the draft has had a chance to save, causing them to lose their draft.

This fixes issue #4985.