issue: Support Exchange Shared Mailbox Auth
This addresses an issue where the system cannot successfully authenticate to
an Exchange server using shared mailbox auth credentials (eg. Username:
`main-user@domain.com\shared-mailbox@domain.com`). The system passes the
entire Username value as the `username` parameter for `imap_open()` and
Exchange fails to authenticate. This adds two new variables called
`authuser` and `username` to class MailFetcher. This also updates the
`MailFetcher::__construct()` method to explode the Username value at the
backslash giving us two parts: the Authentication Username and the Mailbox
Username. If there are in fact two parts we will assign the first part to
the new `authuser` variable and assign the second part to the new `username`
variable. If we detect an `authuser` we will add it to the `imap_open()`
flag called `authuser` and append it to the `srvstr`. In addition this
updates the `MailFetcher::getUsername()` method to return the new `username`
variable instead of the hash table data.