feature: Email Ticket Export To Staff
This adds a new feature where any Agent clicking Export on any Ticket Queue
will send a request to the backend where we create a temporary file and add
the ticket export data. At that point we will check to see if the csv build
took longer than the `max_execution_time` to build the csv and if the build
took less time than the `max_execution_time` then we download in browser. If
the csv build takes longer than the `max_execution_time` then we email it to
whomever initiated the Export. This should cut down/stop the export timeouts
with large datasets. The export timeout issues were caused by the system
starting the download and trying to build the csv file while it’s
downloading. The other cause/factor in the timeouts was the ticket queue
page continuing to load (meaning the request is still open) until the export
completely finished downloading. So all-in-all you had the page loading
(waiting for export completion), a download starting, and a file being built
whilst downloading all in the same request. This feature makes it to where
if an Agent clicks Export it sends an AJAX POST request in the background,
the ticket queue page fully loads instantly and waits for nothing, the csv
file is built in the background using tmp files, and once complete the file
is then either passed to the browser for download or passed to the mailer to
be mailed out accordingly.

In addition this adds reusable JS functions to trigger a csv build, trigger
automatic csv check status, and trigger the finalization of the export
(download route or email route).