issue: Default Queue Sorting
This addresses an issue where Queue Exports do not have a default sort and
therefore are sorted incorrectly by ASC (old tickets on top latest tickets
on bottom). This adds a new method to class Queue called `getColumn()` that
takes a Column ID and returns the appropriate column (useful for applying
specific sorts to a specific column). In addition this updates the
`getSortOptions()` method to check for sorts saved in the session. If sorts
for the queue are saved in the SESSION we will apply them to the Queue, if
none however we will use the Default Sort.