Delete Threads/Entries of Deleted Tickets
This commit reallows us to delete threads and thread entries when deleting tickets by adding a column called thread_type to the thread event table that contains the same data as the object_type field on the thread table. The reason for doing this was because in the dashboard reports, we were relying on the object type in the thread table to give us accurate report data.