tasks: Add feature to cascade cancellation
So, when a task is cancelled, offer the option to also cancel all
dependent tasks, and all tasks dependent on those tasks. The list is
presented in the dialog as a checkbox list, and, by default, all tasks
are selected to be cancelled.

Cancelling only some dependent tasks currently has the caveat that tasks
which depend on a cancelled task might not be started when they are able
to be started. In such a case, an agent can manually start or cancel the
task at a later date.