Modify Draft Saving
This commit adds a button to redactor to save drafts automatically when the admin clicks it, rather than having to wait 30 seconds every time.

Note: This option currently only appears if there is an existing draft already

Also, it will automatically save drafts on tickets if a field that causes the page to refresh is clicked. These drafts save regardless of if there is an existing draft. Those fields are:

- Status
- Department
- User (includes Change Owner in dropdown)