issue: Form Field Flags
This addresses an issue where any Agent can change the Type and Variable of
some of the core fields. In most cases, changing the Type and Variable of
core fields will cause issues later on down the line. With this in mind, we
need to set the flags to where Agents cannot change the Type and Variable of
specific core fields to avoid any issues. This updates the installer to
reflect the additional flags and updates the Upgrader to add these flags to
existing fields.